Remove Camera Capture Popup
