+++ stuff/config/qss/Medium/less/layouts/misc.less
@@ -2,12 +2,6 @@
    Misc 
 ----------------------------------------------------------------------------- */
 
-PencilTestPopup {
-  min-height: 730px; /* Allow for using a 768 screen */
-  min-width: 512px;  /* some clipping will still occur on width, but this
-                        allows for filling half of a 1024 screen */
-}
-
 // #HistoryPanel {}
 
 #MatchLineButton {
