+++ stuff/config/qss/Medium/less/layouts/popups.less
@@ -88,10 +88,8 @@ ProjectPopup {
 }
 
 /* -----------------------------------------------------------------------------
-   PencilTestPopup / CameraCapture Window
+   CameraCapture Window
 ----------------------------------------------------------------------------- */
-
-//PencilTestPopup {}
 
 #GearButton {
   &::menu-indicator{
