+++ stuff/config/qss/Neutral/Neutral.qss
@@ -1437,13 +1437,6 @@ ParamsPage {
 /* -----------------------------------------------------------------------------
    Misc 
 ----------------------------------------------------------------------------- */
-PencilTestPopup {
-  min-height: 730px;
-  /* Allow for using a 768 screen */
-  min-width: 512px;
-  /* some clipping will still occur on width, but this
-                        allows for filling half of a 1024 screen */
-}
 #MatchLineButton {
   background-color: #a6a6a6;
 }
@@ -1693,7 +1686,7 @@ ProjectPopup QLabel {
   color: #000000;
 }
 /* -----------------------------------------------------------------------------
-   PencilTestPopup / CameraCapture Window
+   CameraCapture Window
 ----------------------------------------------------------------------------- */
 #GearButton::menu-indicator {
   image: url("");
