+++ stuff/profiles/layouts/rooms/Default/menubar_template.xml
@@ -235,8 +235,6 @@
         <command>MI_OpacityCheck</command>
         <command>MI_CameraTest</command>
         <command>MI_Cleanup</command>
-        <separator/>
-        <command>MI_PencilTest</command>
     </menu>
     <menu title="Render">
         <command>MI_PreviewSettings</command>
