+++ stuff/profiles/layouts/shortcuts/otretas.ini
@@ -253,7 +253,6 @@ MI_PasteNumbers=
 MI_PasteValues=
 MI_Pause=
 MI_PCheck=
-MI_PencilTest=
 MI_PickStyleAreas=
 MI_PickStyleLines=
 MI_Play=Return
