+++ toonz/sources/stopmotion/stopmotion.h
@@ -24,6 +24,7 @@
 #include "stopmotionserial.h"
 #include "stopmotionlight.h"
 
+#include "toonz/namebuilder.h"
 #include "toonz/txshsimplelevel.h"
 
 #include <QObject>
@@ -32,6 +33,21 @@
 class QCamera;
 class QCameraInfo;
 class QTimer;
+
+
+//=============================================================================
+// FlexibleNameCreator
+// Inherits NameCreator, added function for obtaining the previous name and
+// setting the current name.
+
+class FlexibleNameCreator final : public NameCreator {
+public:
+    FlexibleNameCreator() {}
+    std::wstring getPrevious();
+    bool setCurrent(std::wstring name);
+};
+
+//=============================================================================
 
 enum ASPECT_RATIO { FOUR_THREE = 0, THREE_TWO, SIXTEEN_NINE, OTHER_RATIO };
