+++ toonz/sources/stopmotion/stopmotioncontroller.h
@@ -5,7 +5,8 @@
 
 // TnzCore includes
 #include "stopmotion.h"
-#include "penciltestpopup.h"
+#include "toonzqt/dvdialog.h"
+#include "toonzqt/lineedit.h"
 
 // TnzQt includes
 #include "toonzqt/tabbar.h"
@@ -50,6 +51,102 @@ class QToolBar;
 class QTimer;
 class QGroupBox;
 class QPushButton;
+class CameraCaptureLevelControl;
+class QRegExpValidator;
+class QCheckBox;
+
+namespace DVGui {
+    class FileField;
+    class IntField;
+    class IntLineEdit;
+}  // namespace DVGui
+
+
+//=============================================================================
+// FrameNumberLineEdit
+// a special Line Edit which accepts imputting alphabets if the preference
+// option
+// "Show ABC Appendix to the Frame Number in Xsheet Cell" is active.
+//-----------------------------------------------------------------------------
+
+class FrameNumberLineEdit : public DVGui::LineEdit {
+    Q_OBJECT
+        /* having two validators and switch them according to the preferences*/
+        QIntValidator* m_intValidator;
+    QRegExpValidator* m_regexpValidator;
+
+    void updateValidator();
+    QString m_textOnFocusIn;
+
+public:
+    FrameNumberLineEdit(QWidget* parent = 0, int value = 1);
+    ~FrameNumberLineEdit() {}
+
+    /*! Set text in field to \b value. */
+    void setValue(int value);
+    /*! Return an integer with text field value. */
+    int getValue();
+
+protected:
+    /*! If focus is lost and current text value is out of range emit signal
+    \b editingFinished.*/
+    void focusInEvent(QFocusEvent*) override;
+    void focusOutEvent(QFocusEvent*) override;
+    void showEvent(QShowEvent* event) override { updateValidator(); }
+};
+
+//=============================================================================
+
+class LevelNameLineEdit : public QLineEdit {
+    Q_OBJECT
+        QString m_textOnFocusIn;
+
+public:
+    LevelNameLineEdit(QWidget* parent = 0);
+
+protected:
+    void focusInEvent(QFocusEvent* e);
+
+protected slots:
+    void onEditingFinished();
+
+signals:
+    void levelNameEdited();
+};
+
+//=============================================================================
+// PencilTestSaveInFolderPopup
+//-----------------------------------------------------------------------------
+
+class StopMotionSaveInFolderPopup : public DVGui::Dialog {
+    Q_OBJECT
+
+    DVGui::FileField* m_parentFolderField;
+    QLineEdit* m_projectField, * m_episodeField, * m_sequenceField, * m_sceneField,
+        * m_subFolderNameField;
+
+    QCheckBox* m_subFolderCB, * m_autoSubNameCB, * m_createSceneInFolderCB;
+    QComboBox* m_subNameFormatCombo;
+
+    void createSceneInFolder();
+
+public:
+    StopMotionSaveInFolderPopup(QWidget* parent = 0);
+    QString getPath();
+    QString getParentPath();
+    void updateParentFolder();
+
+protected:
+    void showEvent(QShowEvent* event);
+
+protected slots:
+    void updateSubFolderName();
+    void onAutoSubNameCBClicked(bool);
+    void onShowPopupOnLaunchCBClicked(bool);
+    void onCreateSceneInFolderCBClicked(bool);
+    void onSetAsDefaultBtnPressed();
+    void onOkPressed();
+};
 
 //=============================================================================
 // StopMotionController
@@ -94,19 +191,21 @@ class StopMotionController final : public QWidget {
       *m_webcamExposureSlider, *m_webcamBrightnessSlider,
       *m_webcamContrastSlider, *m_webcamGainSlider, *m_webcamSaturationSlider,
       *m_liveViewCompensationSlider;
-  QComboBox *m_cameraListCombo, *m_exposureCombo, *m_fileTypeCombo,
-      *m_whiteBalanceCombo, *m_resolutionCombo, *m_imageQualityCombo,
-      *m_pictureStyleCombo, *m_controlDeviceCombo, *m_captureFramesCombo;
+  QComboBox* m_cameraListCombo, * m_exposureCombo, * m_fileTypeCombo,
+      * m_whiteBalanceCombo, * m_resolutionCombo, * m_imageQualityCombo,
+      * m_pictureStyleCombo, * m_controlDeviceCombo, * m_captureFramesCombo,
+      * m_colorTypeCombo;
   LevelNameLineEdit *m_levelNameEdit;
   QCheckBox *m_blackScreenForCapture, *m_placeOnXSheetCB, *m_directShowCB,
       *m_liveViewOnAllFramesCB, *m_useMjpgCB, *m_useNumpadCB, *m_drawBeneathCB,
-      *m_timerCB, *m_showScene1, *m_showScene2, *m_showScene3;
+      *m_timerCB, *m_showScene1, *m_showScene2, *m_showScene3; //, *m_upsideDownCB;
+  CameraCaptureLevelControl* m_camCapLevelControl;
   DVGui::FileField *m_saveInFileFld;
   DVGui::IntLineEdit *m_xSheetFrameNumberEdit;
   FrameNumberLineEdit *m_frameNumberEdit;
   DVGui::IntField *m_onionOpacityFld, *m_postCaptureReviewFld,
       *m_subsamplingFld;
-  PencilTestSaveInFolderPopup *m_saveInFolderPopup;
+  StopMotionSaveInFolderPopup *m_saveInFolderPopup;
   DVGui::IntField *m_timerIntervalFld;
   DVGui::ColorField *m_screen1ColorFld, *m_screen2ColorFld, *m_screen3ColorFld;
   QGroupBox *m_screen1Box;
@@ -290,6 +389,8 @@ protected slots:
   void onWebcamGainSliderChanged(int value);
   void onWebcamSaturationSliderChanged(int value);
   void getWebcamStatus();
+  void onColorTypeComboChanged(int index);
+  void onUpdateHistogramCalled(cv::Mat img);
 
   void onTakeTestButtonClicked();
   void onRefreshTests();
