+++ toonz/sources/stopmotion/webcam.cpp
@@ -24,6 +24,7 @@ TEnv::IntVar StopMotionUseMjpg("StopMotionUseMjpg", 1);
 Webcam::Webcam() {
   m_useDirectShow = StopMotionUseDirectShow;
   m_useMjpg       = StopMotionUseMjpg;
+  m_lut = cv::Mat(1, 256, CV_8U);
 }
 
 //-----------------------------------------------------------------
@@ -127,6 +128,24 @@ bool Webcam::getWebcamImage(TRaster32P& tempImage) {
   if (!error) {
     cv::cvtColor(imgOriginal, imgCorrected, cv::COLOR_BGR2BGRA);
     cv::flip(imgCorrected, imgCorrected, 0);
+    emit updateHistogram(imgCorrected);
+
+    bool convertBack = false;
+    if (m_colorMode != 0) {
+        cv::cvtColor(imgCorrected, imgCorrected, cv::COLOR_RGB2GRAY);
+        convertBack = true;
+    }
+
+    // color and grayscale mode
+    if (m_colorMode != 2)
+        adjustLevel(imgCorrected);
+    else
+        binarize(imgCorrected);
+
+    if (convertBack) {
+        cv::cvtColor(imgCorrected, imgCorrected, cv::COLOR_GRAY2BGRA);
+    }
+
     int width  = m_cvWebcam.get(3);
     int height = m_cvWebcam.get(4);
     int size   = imgCorrected.total() * imgCorrected.elemSize();
@@ -426,6 +445,43 @@ void Webcam::openSettingsWindow() {
     }
   }
   m_cvWebcam.set(cv::CAP_PROP_SETTINGS, 1.0);
+}
+
+//-----------------------------------------------------------------
+
+void Webcam::adjustLevel(cv::Mat& image) {
+    if (m_black == 0 && m_white == 255 && m_gamma == 1.0) return;
+
+    cv::LUT(image, m_lut, image);
+}
+
+//-----------------------------------------------------------------
+
+void Webcam::binarize(cv::Mat& image) {
+    cv::threshold(image, image, m_threshold, 255, cv::THRESH_BINARY);
+}
+
+//-----------------------------------------------------------------
+
+void Webcam::computeLut() {
+
+    const float maxChannelValueF = 255.0f;
+
+    float value;
+
+    uchar* p = m_lut.data;
+    for (int i = 0; i < 256; i++) {
+        if (i <= m_black)
+            value = 0.0f;
+        else if (i >= m_white)
+            value = 1.0f;
+        else {
+            value = (float)(i - m_black) / (float)(m_white - m_black);
+            value = std::pow(value, 1.0f / m_gamma);
+        }
+
+        p[i] = (uchar)std::floor(value * maxChannelValueF);
+    }
 }
 
 //-----------------------------------------------------------------
