+++ toonz/sources/toonz/CMakeLists.txt
@@ -129,7 +129,6 @@ set(MOC_HEADERS
     ../stopmotion/stopmotionserial.h
     ../stopmotion/stopmotionlight.h
     cameracapturelevelcontrol.h
-    penciltestpopup.h
 )
 
 set(HEADERS
@@ -359,7 +358,6 @@ set(SOURCES
     ../stopmotion/stopmotionserial.cpp
     ../stopmotion/stopmotionlight.cpp
 	cameracapturelevelcontrol.cpp
-    penciltestpopup.cpp
 )
 
 add_translation(toonz ${HEADERS} ${SOURCES})
