+++ toonz/sources/toonz/menubar.cpp
@@ -633,8 +633,6 @@ QMenuBar *StackedMenuBar::createFullMenuBar() {
   addMenuItem(scanCleanupMenu, MI_CleanupPreview);
   addMenuItem(scanCleanupMenu, MI_CameraTest);
   addMenuItem(scanCleanupMenu, MI_Cleanup);
-  scanCleanupMenu->addSeparator();
-  addMenuItem(scanCleanupMenu, MI_PencilTest);
 
   // Menu' VIEW
   QMenu *viewMenu = addMenu(tr("View"), fullMenuBar);
