+++ toonz/sources/toonz/menubarcommandids.h
@@ -411,7 +411,6 @@
 
 #define MI_About "MI_About"
 #define MI_StartupPopup "MI_StartupPopup"
-#define MI_PencilTest "MI_PencilTest"
 #define MI_AudioRecording "MI_AudioRecording"
 #define MI_LipSyncPopup "MI_LipSyncPopup"
 #define MI_AutoInputCellNumber "MI_AutoInputCellNumber"
